Simplify MPI interface and fix submodule file name.
The MPI interface to the subroutines that create and cleanup the MPI environment was simplified to hide and check the ierr code inside the submodule. Also fixed the submodule file name to be parallel_mpi.f90 (rather than mpi_parallel.f90).

Finally, added a Makefile to build the example using Intel MPI and ifort.

Still to do: call mpi_abort if there is an error after mpi_init and mpi_finalize.