Add error handling functionality
Added a new subroutine "error_handling" which, when using MPI, will use the ierr code and print an error message if the error message indicates that the MPI operation was not successful. The error_handling routine can be called after every call to MPI.

Also changed the names of the files that implement the parallel functionality with MPI to use "_mpi" as a final suffix, i.e. parallel_mpi_env.f90 becomes parallel_env_mpi.f90 etc.