Make sure the type kinds of reduction operators match.
When using MPI, the reduction operators are set to the MPI constants MPI_SUM, MPI_MPI, etc. In order to make sure ther is no kind mismatch, the type definition has been updated to set the integer kind of the reduction operator components, and when the MPI constant values are assigned, they are set to that same kind.

For now, (kind=4) has been used, this can be changed to be an ASiMoV-CCS specific kind (for example, accs_int or accs_mpiint).