Create CAF version of parallel Pi example
In order the create a pure CAF version of the parallel Pi example code, the core computation part of the code has to be separated out. A module compute_mod.f90 contains the interface for the subroutine compute_pi, which is then implemented using either MPI or CAF in two submodules. Support for CAF was added with a new parallel environment and utilities (sync and timer).

The Pi main code is near identical for both versions - the only difference is in the initial parallel environment that is created.