Clean up some associate code
In accsmat_petsc/set_matrix_values() might as well associate the
members of mat_values before the select type on the matrix as
mat_values is a fixed type.

In accs_solver_petsc we can clean up the deconstruction of the
solver object (whose members need to be type-selected) by performing
the associate at the same time as type selection.