CAF example of paralle protoye fails to link
Ttype bound procedures cannot be implemented in a submodule. Although the MPI version of the parallel prototype still builds, the CAF version does not link because the `set_mpi_reduction_operator` symbol is undefined. The solution is to put the implemenation of the type-bound procedure into the same module where the type is defined - otherwise in instances where the submodule is not linked, the definition of the type is incomplete causing the linker to fail even when the type is not used.