`intent(out)` dummy argument is not explicitly assigned a value
The `result_value` `intent(out)` argument is not explicitly assigned a value - the value comes from the call to MPI_Allreduce, however if those calls fail for whatever reason, the argument is undefined. The Intel compiler picks up on this, gfortran and Cray do not.
It is not possible to assign a default value as part of `class default` - the Intel compiler throws the following error:

```
parallel_collectives_mpi.f90(41): error #8304: In an intrinsic assignment statement, variable shall not be a non-allocatable polymorphic.   [RESULT_VALUE]
            result_value = -99
```
The solution is to mark `result_value` as `intent(inout)`.