Remove coarray compile flag from files that do not need it
Adding the coarray compile flag to all files when using the Intel compiler causes the MPI initialisation to fail and as a result the program fails silently. The interoperability between MPI and CAF is implementation dependent, and in order to avoid this issue it is important to only compile CAF files with the CAF compiler flag.