Integrate `parallel_environment` into Poisson solver
Remove direct references to MPI and PETSc from `poisson.f90` by integrating the code with the `parallel_environment`. This involved changing the signatures of the `create_vector`, `create_matrix` and `crate_solver` subroutines - they now take a `par_env` as an argument. This allows the user to control which parallel environment to create their objects on (if there are more than 1).

Similar, `create_mesh` also takes `par_env` as an argument.