Change calculation of transport equation coefficients
All coefficients are now computed in the form

aP phiP + aF phiF = b

which helps eliminate juggling of signs.

Also changes how boundary conditions are computed - a new subroutine
is added to compute the ghost value phiF in terms of boundary
coefficients, i.e.

phiF = a phiP + b

which can either be evaluated explicitly or used in the equation
coefficients.