Rewrite reorder in terms of to(from)
PETSc is returning the old indices in the new ordering, i.e. from(to).
For most operations (in my mind) it makes most sense to know where something is going,
i.e. to(from).

This is easy to rearrange as

  to(from(i)) = i

Appears to converge sensible on 4 ranks with locally reduced mesh bandwidths