Tidy up use of print/compute bandwidth
The computation was in reordering module, printing was in mesh_utils
module - one depends on the other so they belong in the same module
(I picked reordering, but could go in either).

Simplified their use - the high-level user only needs to call the
printing subroutine, it then checks whether it should run or not,
and calls the compute bandwidth subroutine as appropriate.